create shared component for  edit & create event
